backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
qdrant-client==1.13.1
groq==0.14.0
fastembed==0.4.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.20
httpx==0.28.1
